the player login is called twice.. so on second call you have already entered the map.. many errors are result.
the resetTime in table creature.instance is not updated. mostly after 2 hour the instance is deleted from DB, while you are playing inside..  if you then leave the instance and try to re-enter then you are bind to a new instance..
ToDo: you are in bot-group and enter the instance. sometimes the core save you instance as single player instance, not as group.