fix resting..
on many Tavern the resting zzz was not working.
and when enter a chair, then the zzz removes again..

note from script:
on enter tavern, the core are waiting for client sending a CMSG_AREA_CHANGE Opcode. but he dosn't in goldshire. why??
sniffing show, that if player comes from a zone and enter the tavern, the client send CMSG_AREA_CHANGE... fine... But:
when player comes from an area, as in goldshire, the client send a CMSG_ZONE_CHANGE. (and our core set none RESTING_FLAGS)