next try to fix transport update message to player.
problem is: the transport is itself a map.
the transport-gameobject is included in object to map visible check, but not included in a visible check of map to map
so we have to check it separate..
until now this check is done every tick, and needed or not, the complete transport update list are written and send to player. this is a huge part of transmission to each player.
now i have added a unsorted_map to store the last visible state. now we can reduce the sended packet's to only needed Data.